!66 适配 CherryUSB 1.14.0, 支持 USB2 Host/Device, XHCI 和 PCIe-XHCI，重载 configSTACK_DEPTH_TYPE 类型

+++ example/freertos_feature/software_timer/main.c
@@ -55,7 +55,7 @@ int main(void)
     /* used in no-letter-shell mode */
     ret = xTaskCreate((TaskFunction_t)SoftwareTimerExampleTaskEntry,
                       (const char *)"SoftwareTimerExampleTaskEntry",
-                      (uint16_t)4096,
+                      4096,
                       NULL,
                       (UBaseType_t)SOFTWARE_TIMER_EXAMPLE_TASK_PRIORITY,
                       NULL);
