+++ example/freertos_feature/task/main.c
@@ -45,7 +45,7 @@ int main(void)
 #else
     ret = xTaskCreate((TaskFunction_t)TasksExampleEntry,
                      (const char *)"TasksExampleEntry",
-                     (uint16_t)4096,
+                     4096,
                      NULL,
                      (UBaseType_t)2,
                      NULL);
