+++ example/peripheral/can/canfd/src/canfd_id_filter_example.c
@@ -450,7 +450,7 @@ BaseType_t FFreeRTOSCanfdCreateFilterTestTask(void)
     /* canfd filter test1 task */
     xReturn = xTaskCreate((TaskFunction_t)FFreeRTOSCanfdFilterTask,  /* 任务入口函数 */
                           (const char *)"FFreeRTOSCanfdFilterMode1Task",/* 任务名字 */
-                          (uint16_t)1024,  /* 任务栈大小 */
+                          1024,  /* 任务栈大小 */
                           (void *)CAN_FILTER_MODE_1,/* 任务入口函数参数 */
                           (UBaseType_t)CAN_FILTER_TASK_PRIORITY,  /* 任务的优先级 */
                           NULL); /* 任务控制 */
@@ -470,7 +470,7 @@ BaseType_t FFreeRTOSCanfdCreateFilterTestTask(void)
     /* canfd filter test2 task */
     xReturn = xTaskCreate((TaskFunction_t)FFreeRTOSCanfdFilterTask,  /* 任务入口函数 */
                           (const char *)"FFreeRTOSCanfdFilterMode2Task",/* 任务名字 */
-                          (uint16_t)1024,  /* 任务栈大小 */
+                          1024,  /* 任务栈大小 */
                           (void *)CAN_FILTER_MODE_2,/* 任务入口函数参数 */
                           (UBaseType_t)CAN_FILTER_TASK_PRIORITY,  /* 任务的优先级 */
                           NULL); /* 任务控制 */
