+++ example/peripheral/qspi/src/qspi_flash_polled_example.c
@@ -220,7 +220,7 @@ BaseType_t FFreeRTOSQspiPolledTaskCreate(void)
     taskENTER_CRITICAL();                                          /*进入临界区*/
     xReturn = xTaskCreate((TaskFunction_t)FFreeRTOSQspiPolledTask, /* 任务入口函数 */
                           (const char *)"FFreeRTOSQspiPolledTask", /* 任务名字 */
-                          (uint16_t)1024,                          /* 任务栈大小 */
+                          1024,                          /* 任务栈大小 */
                           NULL,                                       /* 任务入口函数参数 */
                           (UBaseType_t)QSPI_POLLED_TEST_TASK_PRIORITY,   /* 任务的优先级 */
                           NULL);                                   /* 任务控制 */
