+++ example/storage/fatfs/main.c
@@ -55,7 +55,7 @@ int main(void)
 #else
     ret = xTaskCreate((TaskFunction_t)FatfsExampleEntry,
                         (const char *)"FatfsExampleEntry",
-                        (uint16_t)4096,
+                        4096,
                         NULL,
                         (UBaseType_t)2,
                         NULL);
